Improve Triple (#60)
* [NFC] Reorganize Triple.swift

This commit moves the enums for triple components into their own extensions and moves their `parse` functions into them as static methods. These enums are so big that this will look like a huge change, but it’s actually NFC.

* [NFC] Remove indices from triple component parse methods

* Add interpolation utility

I’ll be using this shortly.

* [NFC] Drop .unknown triple cases

Use optionals instead of .unknown to represent an unparsed Arch, OS, Vendor, Environment, or ObjectFormat. This cleans up some logic and will clean up even more later on.

* Test that driver infers default triple correctly

* [NFC] Additional Triple refactoring

* Split environment inference from straight-up parsing
* Extract a helper to reduce bounds checking of components

* [NFC] Add structure to triple parsing

* [NFC] Reify types for triple parsing

* Add triple normalization

* [NFC] Factor triple parsing mechanics into its own type

* Add triple normalization tests

* Improve triple test output

* Clean up and improve fidelity of triple normalization

* Improve various Triple members

* Disable nonfunctioning triple test

* Port a ton of triple tests from LLVM

* Enable triple normalization

* [NFC] Delete unused helper

* Restructure Darwin platform-specific code

Redesigns DarwinPlatform to model device/simulator as related cases, then replaces several instances of if/else spaghetti with that.

* Simplify uses of OS predicates

* Improve handling of Darwin versioning

* Add `Triple.version(for:)` method to look up the version matching whatever Darwin platform we are currently using.
* Add `Triple.supports(_:)` and `Triple.FeatureAvailability` to simplify multi-platform version checks

* Rework requiresRPathForSwiftInTheOS code

It was incorrectly placed in a shared location instead of DarwinToolchain, and we can separate the logic from the implementation a little more clearly.

* Switch back to String for Triple name accessors

* Discourage use of raw version accessors

* DarwinPlatform.DeviceOrSimulator -> .Environment

* Delete backwards compatibility code for 2015 clang

* Internal-only _isSimulatorEnvironment

* Test DarwinPlatform and Triple.FeatureAvailability

* [NFC] Avoid testable