Add environment variable to override xcrun find
If an environment variable of the form SWIFT_DRIVER_${TOOL}_EXEC exists, it will now override the result of “xcrun —find”. For instance, SWIFT_DRIVER_SWIFT_EXEC can be used to change the path used for the swiftCompiler tool. This means you can use a Swift frontend other than the one in your currently selected toolchain.