ADD - look for tools relative to the executable path.
Once swift-driver is installed in the Toolchain, we can look for tools right next to it's executable path, this help us to be more platform agnostic while locating the toolchain.

This is the first step in this direction, we will need to make this more robust to get rid of xcrun if possible, and get a proper linux implementation.