[WIP] Invoke the Swift frontend to prescan for dependencies.
Add a new flag -driver-prebuild-module-dependencies to instruct the driver
to have the Swift frontend prescan for dependencies, which will allow the
driver to explicitly schedule builds for both Swift and C modules,
eliminating the need for the frontend to rebuild those modules when it
tries to import them. This allows for an overall more coordinated
build.

At present, this WIP invokes the Swift frontend and then parses the
JSON before dumping out the dependency graph, so it's not actually
useful for anything.