[WIP] Print commands required to explicitly build module dependencies. This PR resurrects #47 and builds on top of it.
Adds new flag -driver-print-module-dependencies-jobs to instruct the driver to scan for dependencies and generate build jobs for each dependent Swift or C module.

At present, this WIP does the following:
- Invoke the Swift frontend Fast Dependency Scanner to prescan for dependencies.
- Construct a Module Dependency Graph from the dependency scanner's JSON output.
- Generate and print build jbos for each dependent module.

Next Steps:
Actually have the driver schedule and execute the dependent module build jobs ahead of time, eliminating the need to re-build those modules when it tries to import them.