Significantly improve prefix trie performance and memory usage
The previous prefix trie made a new `Node` for each character along the
input string. In addition to being wasteful, this would introduce many
more hops to match a given string.

Instead, store common `Substring`s in the `Node`s, and use a sorted
array/binary search through children to avoid extra time spend hashing
and storing in hashed collections.

Doing this reduces the number of nodes in the trie for the driver's
options by an order of magnitude -- from 6601 nodes to 660.

(This is actually a port of LLVM's Trie, but cleaned up and made more
idiomatic in Swift)