[Explicit Module Builds] Support Versioned PCMs
There exists a strict limitation that a compatible PCM must have the exact same target version with the loading module (either a Swift module or another PCM). This means that each PCM may have to be built multiple times, once for each unique (SDK, architecture, toolchain) tuple of some loading module.

This PR introduces support for generating versioned PCMs for each loading module.
Because now there is more state that we needed to keep track of when generating jobs for module dependencies, this state, along with the Inter-Module Dependency Graph, now live in the `ExplicitModuleBuildHandler`, which is constructed on-demand by the driver in the explicit module build mode. The handler is responsible for generating dependency build jobs and for vending out dependency modules as inputs/command-line options whenever the driver creates a compile job in Explicit Module Build Mode.

Because this PR rewrites/refactores the majority of existing Explicit Module Build code, I tooke the opportunity to re-organize some of the files as well.

Resolves rdar://problem/63944640