Use the Swift frontend to determine the host target triple.
Introduce a utility function to invoke the Swift frontend with
its -print-target-info option and parse the resulting JSON, which
provides information about the target of the compile that will be
useful to the driver.

For now, use this to retrieve the host target triple when no triple
was supplied. This is more consistent than going through Clang to
retrieve this information, and we'll be using more of this info soon.