Use frontend's -print-target-info to determine runtime compatibility info.
Early in driver construction, use -print-target-info to extract information
about the target from the frontend. Use it to produce canonical target and
target-variant triples. Fix up a few tests that weren't updated to
account for canonicalization adding "-simulator" for Intel
iOS/tvOS/watchOS targets.

Then, replace the ad hoc logic that determines the Swift runtime
compatibilty version with the version reported by the frontend. This
is guaranteed to stay up-to-date with the Swift frontend.