[Explicit Module Builds] Avoid adding duplicate module dependency options to compile commands.
When traversing transitive dependencies of some module, there may be many paths that encountered the same module.
We currently ensure we only generate a compile job for each module once, but do not ensure each dependency gets
added as an input to the build command once. This PR fixes that.