[Explicit Module Builds] Refactor the InterModuleDependencyGraph to distinguish prebuilt external dependencies. This is a refactoring change intended to reduce ambiguity in what nodes in the dependency graph can represent. As dependency graph structure and individual modules are being communicated and merged among different components, such ambiguities in the graph introduce additional, unnecessary complexity to graph operations.
- Introduce a new module kind: `.swiftPrebuiltExternalDependency`.
This node kind is to be used for external target dependencies built by build system clients such as SwiftPM. Previously we were using `.swift` modules to represent both pre-compiled external targets and Swift module dependencies that require a build job to be created by the driver. This new structure means that every `.swift` module dependency must have a build job constructed for it, and every `.swiftPrebuiltExternal` module dependency does not require to be built.
- Factor out and unify common graph operations.