Add -emit-module-separately which doesn't use the merge-module phase
Incremental compilation currently relies on a merge-module phase. A typical
build plan for a module App composed of two source files (A.swift and B.swift)
involves four main jobs:

1. Compile A.swift, output A.o and A~partial.swiftmodule.
2. Compile B.swift, output B.o and B~partial.swiftmodule.
3. Link A.o and B.o in the final executable.
4. Merge-module: merging A~partial.swiftmodule and B~partial.swiftmodule to form
the final App.swiftmodule.

Jobs 3 and 4 must run after jobs 1 and 2 to use their outputs.

The flag -experimental-emit-module-separately replaces the merge-module job but
the emit-module job that can run in parallel to the other jobs and doesn’t need
the partial.swiftmodules.

1. Compile A.swift, output A.o.
2. Compile B.swift, output B.o.
3. Link A.o and B.o in the final executable. (only job 3 must run after jobs 1.
   and 2.)
4. Emit-module: Parse A.swift and B.swift to generate the final App.swiftmodule.

Early comparison with merge-module indicates similar performances and even a
slight performance gain with all of the optimizations applied here. However,
the use of `-experimental-skip-non-inlinable-function-bodies` may be too
aggressive.

rdar://69330639