[ExplicitModule] when using driver as an executable, differentiate PCM dependency names for each target
We need this to enable explict modules in the driver-as-executable mode. For instance,
we have two Swift targets A and B, where A depends on X.pcm which in turn depends on Y.pcm, and
B only depends on Y.pcm. In the driver-as-executable mode, the build system isn't aware
of the shared dependency of Y.pcm so it will be generated multiple times. If all these Y.pcm
share the same name, X.pcm may fail to be loaded because its dependency Y.pcm may have a changed mod time.

We only differentiate these PCM names in the non-integrated mode due to the lacking of
inter-module planning.

rdar://72830153