Add Round-Trip Serialization For ModuleDependencyGraph
Use bitstream to read and write ModuleDependencyGraph instances out to
disk. The idea is that eventually the Driver can emit its dependency
graph as a supplemental output (possibly adjacent to the build record)
and recover the state it was last in by reading that file. This is as
opposed to the status quo where O(N) swiftdeps files are read and
integrated at startup. Whether this is a net win has yet to be seen.

Also, we're using bitstream because it's convenient. It would be prudent
to explore alternative serialization formats at some point.

rdar://73517080