Shuffle Fingerprints Into Serialized DepGraph Nodes
Place fingerprints as a trailing blob of bytes at the end of a module
dep graph node record instead of as a dedicated record. This allows us
to tighten up the deserialization path a touch. Note that we aren't
adding fingerprints to the strings table because collisions are expected
to be extremely rare within the same graph - there would be very little
sharing to be had.