Serialize The Uses-By-Defs Map
This completes the serialization of the dependency graph.

The idea is to notice that the dependency key for a def may not always appear in the node map, but its uses certainly will (there are scattered invariants, especially in the tracer, that this is true). That means we'll be redundantly serializing nodes if we just naively write out the contents of the uses-by-defs map. Therefore, we establish a linear ordering of dependency nodes - in this case, serialization order - and record memoized node IDs. These are recorded instead, saving 6x the space of a regular serialized node.