[Integration Tests] Calculate the swiftRootDir more accurately
Previously the swiftRootDir would be calculated by taking
swiftBuildDir/../../.., which is accurate on most developer machines,
but it is not accurate for some CI bots that use one fewer layer of
directories. We already know the swiftRootDir more accurately, we're
using it for calculating relative paths, just use that one.