Do not add precompiled bridging header to a Compile job's display inputs "Display inputs" are used to convey to the user/client the set of input files that a given compile job will build, e.g. in parseable-output.
The driver is relying on the presence of the bridging PCH in `displayInputs` because the `printActions` code relies on `displayInputs`, when they are present. Unlike parseable-output, action printing is used for compiler debugging purposes only, and given that `printActions` is a best-effort emulation of the legacy driver behavior that doesn't really exist in the new driver, we can just as well get away with using `primaryInputs` and ensure `compile` job actions refer to the corresponding precompiled header, to keep the C++ tests passing (`test/Driver/actions.swift` & `test/Driver/actions-dsym.swift`).