Perform Input Modification Checks At the End of the Build
Also deduplicate them while we're at it. We don't need to ask each job for its input set, we just need to check the same set of files we recorded mod times for in the first place. This saves a quadratic amount of disk I/O in the incremental and non-batch builds, which can be a huge wall time savings for large compiles.