Use a Confinement Queue to Protect IncrementalCompilationState
This class is serendipitously used only from the main queue for the
Driver's usual use cases. But we have no reason to expect clients of
libSwiftDriver to manipulate this class in the same manner. Provide a serial
confinement queue and protect the mutable members of
IncrementalCompilationState with it.

In the future, we ought to drop the queue and use actors.

rdar://75744005