Use unique temporary file names This PR augments `VirtualPath` with an ability to create a temporary file that is unique in its execution context: `createUniqueTemporaryFile`. i.e. if a temporary file with the same filename as one being requested has been created before, using this API, then the newly-created file will have a suffix appended to the base name to distinguish it.
This brings swift-driver closer to the legacy driver's behavior, which generates unique temporary files by default, using LLVM's `llvm::sys::fs::createTemporaryFile`.
Doing so is necessary to resolve several kinds of problems, for example: compiling a module `Foo` consisting of two source files `Foo.swift` and `Bar.swift` will result in a temporary partial module `Foo.swiftmodule` being produced from building `Foo.swift` and a full `Foo.swiftmodule` produced from the merge-modules job. If the full module does not need to be exposed to the user as the output of compilation, it will be a temporary file, in which case it will conflict with the partial `Foo.swiftmodule` causing the executor to rightly complain about potential cycles in the build.

I entertained the idea of instead resolving `.temporary` paths to unique files at argument resolution time, but that proved troublesome because in that case a specific unique file has to be tracked via a unique `VirtualPath.handle`, where different such handles may refer to the same `VirtualPath`, which is an opaque distinction to many path clients in the driver, and would be really hard to enforce throughout the code-base. Instead, this approach relies on the file's creator to make sure it is unique.

Having enum case access control would be super useful here in order to prohibit clients from instantiating `.temporary` paths explicitly, instead of via the new API. This is left for future work once Swift has such a feature.

Resolves rdar://76123509