Add workaround for building executables with older CMake versions.
Some time between CMake 3.16 and 3.17 the default behaviour seems to have changed as follows:
For a project which will build an executable, e.g.:
```
add_executable(exe_target
    main.swift)
target_link_libraries(exe_target PUBLIC
    libA
    libB)
```
In 3.17, the invocation command to create the `exe_target` executable includes `-Xlinker -rpath -Xlinkier <libA_path/lib>` and so forth for the linked libraries and their dependencies.
In 3.16, the linker is not passed the rpath flags, so the executable ends up without the `LC_RPATH` required for it to run.
This affects executable targets only, libraries do still have rpaths propagated correctly.