Refactor `IncrementalCompilationState` to compute initial state before job-generation.
In the near future, planning of `PrecompileModuleDependenciesJobs`, for explicit module dependencies, will require access to the incremental state in order to incrementalize expensive actions such as dependency scanning.

This change refactors the creation of incremental compilation state in order to move computation of the initial state (dependency graph) to occur before job-generation.
The first wave of jobs is then computed after job-generation, using the initial state, and the set of jobs in the plan, as input.

- Separate `InitialIncrementalStateComputer` to run early, by the `Driver`, to compute the dependency graph and the set of changed inputs. Rename `InitialIncrementalStateComputer` into `IncrementalDependencyAndInputSetup`.
- Introduce `FirstWaveComputer` which the constructor of the `IncrementalCompilationContext` uses to compute the `mandatoryJobsInOrder` set of jobs the executors *must* run.