Do Not Serialize Ephemeral Input Mappings in Priors
Build systems reserve the right to alter the mapping from swift files to
swiftdeps files at will. We cannot rely on output file maps having the
same entries from run-to-run, so we cannot rely on combining
deserialized inputs from priors and the output file map to agree.

In rare circumstances, XCBuild would alter the GUID suffix it used for
a swiftdeps file in the OFM, and the priors file would enter the old
swiftdeps path into the dependency graph. This would cause a crash when
that bogus file was looked up during integration.

A far more common scenario is the GUID for the file being altered by an
"idempotent" add-remove operation. When the file is re-added, its entry
in the output file map will have shifted which makes the priors
disagree.

The easy solution is to just ignore the mapping nodes here. The only
source of truth for this mapping is the output file map, and serializing
this into the priors is just duplicative and wrong.

This change drops swiftdeps in priors from consideration during
(de)serialization. A subsequent commit will remove them from the priors
format and bump the format.

The proper answer to rdar://77804636