Consume The Non-Owning Bitstream Deserialization IR
Blobs are now decoded as ArraySlices, so we need to switch to the stdlib entrypoint. This has the added upside of being *much* more performant than String.init(data:encoding:) which not only cost a copy to create the Data, but also could not take advantage of stdlib's fast paths for UTF8 buffers as inputs.