build: remove bogus dependency on `swift-driver`
The `SwiftDriverTests` target should not depend on the `swift-driver`
target.  The latter is the executable binary.  This was noticed when
trying to port the swift-driver project to Windows.  The Windows linker
caught the fact that the `main` entry point was multiply defined, once
from the test binary, and once from the `swift-driver` executable.  Do
not add the `swift-driver` to the dependency set as that will cause a
link dependency, which results in the multiply defined symbols.