SwiftDriver: initial (incomplete) windows port
This is sufficient to build on Windows, and the resulting driver is able
to somewhat function.  The tests are still not usable due to assumptions
about path which do not hold and the insufficiency of the path
representation in tools-support-core.  This is likely best approached by
replacing the use of `Path` from t-s-c with the representation in
system.  However, this allows some attempt to validate the swift-driver
which is increasingly becoming required for s-p-m, and thus unblocks
further work.