SwiftDriver: unwrap swift frontend names
We need to specify both the windows and unix names for the frontend and
compiler binaries.  This is required to avoid an infinite recursion
trying to lookup the tools.  This unblocks a large portion of the test
suite for Windows.

Thanks to @artemcm for helping identify the source of the recursion!