Enable emit-module-separately with wmo
When passed the flag -emit-module-separately-wmo, schedule a separate optimized emit-module job to output the module files even when building in whole-module mode (-wmo). In this mode, the classic wmo job generates only the object file and the emit-module job generates the module files concurrently. With enough available cores, this will generate the files needed to unblock building dependencies early, while the classic wmo job is still building the binary file.

The flag should only be passed down to the compiler when the target has dependencies that will be built in the same run, otherwise if there is no dependencies, this mode mostly duplicates work and doesn't provide any gain.

rdar://85428222