Tests: make ParsableMessageTests pass on Windows
This adjusts the test suite to account for the differences on Windows.

We must construct the absolute path to the working directory, which
includes the drive letter.  Furthermore, construct the paths for the
checks as relative to the working directory.  This makes the test more
explicit about it's expectations and makes it support Windows.

Windows does not support signals, even for unexpected termination.
Instead the process exits with a particular error code which encodes the
"exception".  This adjusts the test expectation accordingly.

Adjust the diagnostic expectations which do not match due to the EOL
specifier (CR+LF vs LF).

The ParsableMessageTests now pass on Windows.