Driver: wire through the actual error from OFM loading
The deserialisation process may fail for reasons other than unable to
load the file.  This was occurring on Windows and was giving an
erroneous error message that the file could not be loaded when the issue
was the contents of the file.  This threads that error through the
wrapped error type rather than just dropping the error on the floor.