Tests: explicitly pass a `ld` for Darwin targets
When testing with explicit Darwin targets, we need to explicitly add the
ld path to use for the tests as not all platforms use or have a `ld`
executable.  Stub out the path to ensure that the tests are able to be
executed.  Alternatively, we could remove the Darwin specific triples on
the test.

Additionally, while in the area, add an additional test case for Windows
x86_64 LTO and enable the linker profile argument checks for ELFish
targets on Windows as well.