Tests: make ExplicitModuleBuildTests pass on Windows
This adjusts the driver invocation to add in additional module search
paths for the tests which seem to be required on Windows.  It is
unclear why this is needed for the swift-driver, as the search paths
should be derived from `-sdk`.  Ensuring that the tests pass is the
first step towards making swift-driver viable for Windows though.