Install swift-system into the toolchain in CMake build install action
https://github.com/apple/swift-driver/pull/915 added a CMake dependency on swift-system.
Since then, the CMake build of the driver builds swift-system, but when the `install` action is triggered, the dependency's module and shared library are not installed into the toolchain. As a result, built toolchains are unable to invoke swift-driver due to the missing dylib.

This ensures that much like other driver dependencies (TSC, ArgumentParser), we also install SwiftSystem.

Resolves rdar://86754758