Tests: add shims and stdlib search paths to driver
The Windows toolchain is setup with a split SDK/toolchain and relies on
being able to compose the two.  For some reason, this requires
additional include paths to be passed to the swift-driver implementation
(`-sdk` is sufficient with the C++ driver).  For now, add the additional
include paths to allow the tests to pass.