Windows: support `-tools-directory`
Un-hardcode the explicit target triple for the test.  This allows the
test to run on Windows where there is no `ld` (nor `ld.exe`).  This then
uncovers the fact that the windows toolchain did not honour
`-tools-directory`.  Fixing that uncovers the fact that the lookup for
the tool will fail as the empty file does not get identified as a
binary.  To work around that, we now emit a 97-byte PE binary that is
correctly identified as a binary as the contents.  This will never be
executed and so should be fine to emit on all platforms.  Note that the
identified linker is different between Darwin and non-Darwin as the
behaviour of the driver is different - non-Darwin uses clang, while
Darwin uses `ld` (ld64) directly.

Thanks to @artemcm for the discussion on how to fix this properly.